Move manager classes to a dedicated folder
