+++ core/src/main.py
@@ -4,9 +4,9 @@ import ray
 import torch
 from loguru import logger
 
+from core.src.managers.training_manager import TrainingManager
 from core.src.settings import get_settings
 from core.src.setup import configure_logging
-from core.src.utils.training_manager import TrainingManager
 
 if __name__ == "__main__":
     os.environ["RAY_COLOR_PREFIX"] = "1"
